requirements .txt
kafka-python==2.0.2
pyspark==3.0.3
tweepy==4.4.0
textblob==0.15.3
findspark==1.3.0
pymongo==4.0.1